pyproject.toml
[project]
name = "meeting-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.116.1",
    "load-dotenv>=0.1.0",
    "open-agent>=0.0.0",
    "openai-agent>=0.1.0",
    "pydantic>=1.10.22",
    "uuid>=1.30",
]
